requirements.txt
discord.py==0.14.2
sleekxmpp>=1.3.1